#463 split out `type-example.php` into multiple individual test files
