+++ psalm.xml
@@ -7,6 +7,7 @@
 >
     <projectFiles>
         <directory name="src" />
+        <directory name="tests/static-analysis" />
         <ignoreFiles>
             <directory name="vendor" />
         </ignoreFiles>
